Add support for postgresql
Have added a db package that will be used by any commands and queries
that need to be made into the calendar_api database. Have also added a
Makefile so that it will be easy to create and drop the database
locally. Instructions on how to use the commands have been added to the
README.