Add support for an indefinite number of test users
Have added support to an indefinite number of users in the
`config/env.yaml` file. This involved the creation of new `ToMapList`
function that will convert nested lists in the yaml file into maps that
can be used across the codebase. The `db-seed` command has been
refactored to use this function.