Run teardown and setup before running each test
Have moved the logic from `db-clean` and `db-seed` commands into the
dbHelpers package so that these can be used by the testsuite as well as
by the make commands. These will now be called by the `TestMain`
function so that the database can be cleaned and popluated correctly
before each test that connects to it.