Fixed issues with the PATCH endpoint
The go-pg package doesn't appear to support the structs library which makes
updating specific parameters a little difficult. As a work-around, I've
added a new SetSqlColumns() function to the dbhelpers package which will
generate part of an SQL string that can be run using go-pg's `Exec()`
function instead. I've also added a call to `FindCandidate()` function to
the CandidatesCommands test as though the fact that object returned by
`UpdateCandidate()` was never persisted in the database itself was
missed in the unit tests.