Add support for the multiple tables to db-seed
Have added the time_slots table to the config/env.yaml file and updated
the `make db-seed` command to support the creation of rows for multiple
tables. This involved the creation of the new `InsertConfSql()` function
in the dbHelpers package. Writing tests for this function was difficult
as the use of unordered maps initially created flakey tests. Have
resorted to the use of regexes in the relevant tests. Other tests that
refer to the candidates defined in the environment variables have had to
be altered accordingly as well.