Create associative table
This will keep track of time slots that belong to candidates. Have
decided to do this as I expect interview time slots to be persisted in
the time_slots table as well so this might be the simplest approach.
Have added a CASCADE clause so as to ensure that the relevant rows in
this table are deleted when a row in either candidates of time_slots is
deleted.