Add the new date column to the TimeSlot model
Have added the the column as a string in the TimeSlot model as this
is easier to marshal and unmarshal to and from a JSON than the time.Date
type and Postgres handles a correctly formatted string just fine. Have
had to update the commands and queries that refer to this as well as any
related tests.