Rename the duration column in time_slots
Renaming the `duration` column to `end_time` as it should be easier to
write the SELECT queries with a range of dates rather than with the
duration. I still think that duration is easier for a user to comprehend
so most likely any frontend application would have to take that into
account in the UI. Renaming the column is fairly straight forward at
this stage so no need to drop duration and add a new column.