Add query for finding available interviewers
Have added a SELECT query that will return a list of available time
slots for a given candidate and two given interviewers. This involved
the creation of the InterviewerAvailability serialiser which will be
used to more adequately format the database response.

The tests are a little light at the moment but I'm tired so whatever.